cloud-sdk: Use `usize` for number of `probe` retries
We can avoid the use of casting (for u64), which can in fact fail, using `usize`
for the number of probe retries. The `usize` use here seems a good fit
as it matches the word size of the current in-use CPU architecture.

+++ updatehub/src/cloud_mock.rs
@@ -39,7 +39,7 @@ impl<'a> Client<'a> {
 
     pub(crate) async fn probe(
         &self,
-        _num_retries: u64,
+        _num_retries: usize,
         _firmware: api::FirmwareMetadata<'_>,
     ) -> Result<api::ProbeResponse> {
         RESPONSE_CONFIG.with(|conf| match std::ops::Deref::deref(&conf.borrow()) {
